Resultatsiden henter nå topliste basert på gjennomsnitt med to decimal for stands med minst N stemmer (Definert i Resulatet klassen, men bør flyttes).

+++ DAT109-Prosjekt-Gruppe9/src/no/hvl/dat109/expo/statistics/StandResult.java
@@ -28,8 +28,8 @@ public class StandResult {
         return acc;
     }
 
-    public Integer getWeightedAverage(){
-        return getTotalPoints() / votes.size();
+    public Double getWeightedAverage(){
+        return getTotalPoints().doubleValue() / votes.size();
     }
 
     public List<Vote> getVotes() {
